[SDTODO-2]: Updates 1. Biometrics added 2. Splash Screen 3. AppIcon added
